ContestUnderAudit has mutable status so can be set before auditing. checkContestsCorrectlyFormed, checkContestsWithCvrs can set contest status tabulateVotes(cvrs.iterator()) make BoulderCounty, Colorado audits get checkContestsWithCvrs correct remove writeins from DominionCvrReader cvrs (Boulder) get Charset as correct as possible in BoulderCounty, Colorado
