+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/RiskTestingFn.kt
@@ -1,21 +1,21 @@
 package org.cryptobiotic.rlauxe.core
 
-// NOTE: leave them in this order, as contest status is min rank
+// NOTE: contest status is min rank of assertions
 enum class TestH0Status(val rank: Int, val complete: Boolean, val success: Boolean) {
     InProgress(0,false, false),
 
     // contest status
-    ContestMisformed(1,true, false), // Contest incorrectly formed
-    MinMargin(2,true, false), // margin too small for RLA to efficiently work
-    TooManyPhantoms(3,true, false), // too many phantoms, makes margin < 0
-    FailMaxSamplesAllowed(4,true, false),  // estimated samples greater than maximum samples allowed
-    AuditorRemoved(5,true, false),  // auditor decide to remove it
-    NoLosers(6,true, true),  // no losers, ie ncandidates = nwinners
+    NoLosers(1,true, true),  // no losers, ie ncandidates <= nwinners
+    ContestMisformed(2,true, false), // Contest incorrectly formed
+    MinMargin(3,true, false), // margin too small for RLA to efficiently work
+    TooManyPhantoms(4,true, false), // too many phantoms, makes margin < 0
+    FailMaxSamplesAllowed(5,true, false),  // estimated samples greater than maximum samples allowed
+    AuditorRemoved(6,true, false),  // auditor decide to remove it
 
     // possible returns from RiskTestingFn
     LimitReached(10,false, false),  // cant tell from the number of samples available
     StatRejectNull(11,true, true), // statistical rejection of H0
-    //// only when sampling without replacement all the way close to Nc
+    //// only happens when sampling without replacement all the way close to Nc
     SampleSumRejectNull(12,true, true), // SampleSum > Nc / 2, so we know H0 is false
     AcceptNull(13,true, false), // SampleSum + (all remaining ballots == 1) < Nc / 2, so we know that H0 is true.
 }
