+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/estimate/TestSampleGenerator.kt
@@ -25,7 +25,7 @@ class TestSampleGenerator {
         if (!silent) println(" N=${cvrs.size} theta=$theta withoutReplacement")
 
         // count actual votes
-        val votes: Map<Int, Map<Int, Int>> = tabulateVotes(cvrs) // contest -> candidate -> count
+        val votes: Map<Int, Map<Int, Int>> = tabulateVotes(cvrs.iterator()) // contest -> candidate -> count
         if (!silent && showContests) {
             votes.forEach { key, cands ->
                 println("contest ${key} ")
